+++ frontend/src/components/ApplicationAcceptedPopup/ApplicationAcceptedPopup.module.scss
@@ -7,7 +7,7 @@
   border-radius: 20px;
   background-color: var(--color-white);
   margin-top: 80px;
-  font-family: "Raleway", "Arial", sans-serif;
+  font-family: var(--font-primary);
   display: flex;
   flex-direction: column;
   align-items: center;
