+++ frontend/src/components/CalendarPopup/CalendarPopup.module.scss
@@ -1,19 +1,9 @@
 .title {
   margin-top: 48px;
-  font-family: 'Raleway', 'Arial', sans-serif;
+  font-family: var(--font-primary);
   font-weight: 700;
   font-size: 32px;
   line-height: 135%;
-
-  @media screen and (max-width: 1320px) {
-    font-size: 26px;
-    line-height: 130%;
-    margin-top: 25px;
-  }
-
-  @media screen and (max-width: 743px) {
-    margin: 25px 44px 0 20px;
-  }
 }
 
 .container {
@@ -22,20 +12,6 @@
   gap: 47px;
   justify-content: center;
   margin-top: 28px;
-
-  @media screen and (max-width: 1320px) {
-    grid-template-columns: 398px 1fr;
-    gap: 20px;
-    margin-top: 23px;
-  }
-
-  @media screen and (max-width: 743px) {
-    grid-template-columns: 1fr;
-    grid-template-rows: 472px 1fr;
-    gap: 32px;
-    margin-top: 2px;
-    width: 100%;
-  }
 }
 
 .infoTitle {
@@ -43,14 +19,6 @@
   font-size: 20px;
   line-height: 120%;
   margin-bottom: 28px;
-
-  @media screen and (max-width: 1320px) {
-    line-height: 135%;
-  }
-
-  @media screen and (max-width: 743px) {
-    margin-bottom: 23px;
-  }
 }
 
 .iconContainersWrapper {
@@ -59,15 +27,6 @@
 
 .form {
   width: 100%;
-
-  @media screen and (max-width: 1320px) {
-    width: 278px;
-  }
-
-  @media screen and (max-width: 743px) {
-    margin-top: -10px;
-    width: 320px;
-  }
 }
 
 .notAuthorizedUserContainer {
@@ -81,14 +40,6 @@
   border-radius: 20px;
   margin-bottom: 26px;
   box-sizing: border-box;
-
-  @media screen and (max-width: 1320px) {
-    margin-bottom: 24px;
-  }
-
-  @media screen and (max-width: 743px) {
-    padding: 10px;
-  }
 }
 
 .linkText {
@@ -96,10 +47,6 @@
   font-size: 14px;
   line-height: 140%;
   letter-spacing: 0.04em;
-
-  @media screen and (max-width: 1320px) {
-    line-height: 135%;
-  }
 }
 
 .link {
@@ -113,28 +60,82 @@
 .buttonContainer {
   margin-top: 5px;
   width: 290px;
+}
 
-  @media screen and (max-width: 1320px) {
+@media screen and (max-width: 1320px) {
+  .title {
+    font-size: 26px;
+    line-height: 130%;
+    margin-top: 25px;
+  }
+
+  .container {
+    grid-template-columns: 398px 1fr;
+    gap: 20px;
+    margin-top: 23px;
+  }
+
+  .infoTitle {
+    line-height: 135%;
+  }
+
+  .form {
+    width: 278px;
+  }
+
+  .linkContainer {
+    margin-bottom: 24px;
+  }
+
+  .linkText {
+    line-height: 135%;
+  }
+
+  .buttonContainer {
     margin-top: 0;
     width: 100%;
   }
+}
 
-  @media screen and (max-width: 743px) {
+@media screen and (max-width: 743px) {
+  .title {
+    margin: 25px 44px 0 20px;
+  }
+
+  .container {
+    grid-template-columns: 1fr;
+    grid-template-rows: 472px 1fr;
+    gap: 32px;
+    margin-top: 2px;
+    width: 100%;
+  }
+
+  .infoTitle {
+    margin-bottom: 23px;
+  }
+
+  .form {
+    margin-top: -10px;
+    width: 320px;
+  }
+
+  .linkContainer {
+    padding: 10px;
+  }
+
+  .buttonContainer {
     margin-top: 5px;
     width: 100%;
   }
-}
 
-.submitButton {
-  @media screen and (max-width: 743px) {
+  .submitButton {
     margin-top: 5px;
     width: 100%;
   }
-}
 
-.info {
-  @media screen and (max-width: 743px) {
+  .info {
     padding: 0 20px;
     width: 320px;
   }
 }
+
