+++ frontend/src/components/RegistrationLessonPopup/RegistrationLessonPopup.jsx
@@ -13,6 +13,7 @@ import { setIsApplicationAcceptedPopupOpen, setIsRegistrationLessonPopupOpen } f
 import { useModalClose } from '../../hooks/useModalClose';
 import { Button } from '../Button/Button';
 import { calculateLessonDuration } from '../../utils/functions/timeFunctions';
+import { GENITIVE_MOHTHS } from '../../utils/consts/constants';
 
 const RegistrationLessonPopup = () => {
   const dispatch = useDispatch();
@@ -30,13 +31,9 @@ const RegistrationLessonPopup = () => {
 
   const getFormattedDate = (dateString) => {
     const date = new Date(dateString);
-    const months = [
-      'января', 'февраля', 'марта', 'апреля', 'мая', 'июня',
-      'июля', 'августа', 'сентября', 'октября', 'ноября', 'декабря',
-    ];
 
     const day = date.getDate();
-    const month = months[date.getMonth()];
+    const month = GENITIVE_MOHTHS[date.getMonth()];
     const hours = date.getHours().toString().padStart(2, '0');
     const minutes = date.getMinutes().toString().padStart(2, '0');
