+++ frontend/src/utils/consts/constants.js
@@ -53,3 +53,12 @@ export const SCREEN_WIDTH = {
   tabletScreen: 744,
   mobileScreen: 550,
 };
+
+export const MONTHS = [
+  'Январь', 'Февраль', 'Март', 'Апрель', 'Май', 'Июнь', 'Июль',
+  'Август', 'Сентябрь', 'Октябрь', 'Ноябрь', 'Декабрь'];
+
+export const GENITIVE_MOHTHS = [
+  'января', 'февраля', 'марта', 'апреля', 'мая', 'июня',
+  'июля', 'августа', 'сентября', 'октября', 'ноября', 'декабря',
+];
